Refactor torch utilities
