+++ examples/torch/pearl_metaworld_ml45.py
@@ -4,18 +4,19 @@
 import click
 import metaworld.benchmarks as mwb
 
+from garage import wrap_experiment
 from garage.envs import GarageEnv, normalize
-from garage.experiment import LocalRunner, wrap_experiment
+from garage.experiment import LocalRunner
 from garage.experiment.deterministic import set_seed
 from garage.experiment.task_sampler import EnvPoolSampler
 from garage.sampler import LocalSampler
+from garage.torch import set_gpu_mode
 from garage.torch.algos import PEARL
 from garage.torch.algos.pearl import PEARLWorker
 from garage.torch.embeddings import MLPEncoder
 from garage.torch.policies import ContextConditionedPolicy
 from garage.torch.policies import TanhGaussianMLPPolicy
 from garage.torch.q_functions import ContinuousMLPQFunction
-import garage.torch.utils as tu
 
 
 @click.command()
@@ -148,7 +149,7 @@ def pearl_metaworld_ml45(ctxt=None,
         reward_scale=reward_scale,
     )
 
-    tu.set_gpu_mode(use_gpu, gpu_id=0)
+    set_gpu_mode(use_gpu, gpu_id=0)
     if use_gpu:
         pearl.to()
