+++ tests/garage/torch/algos/test_pearl.py
@@ -21,13 +21,13 @@ from garage.experiment import LocalRunner
 from garage.experiment.deterministic import set_seed
 from garage.experiment.task_sampler import SetTaskSampler
 from garage.sampler import LocalSampler
+from garage.torch import set_gpu_mode
 from garage.torch.algos import PEARL
 from garage.torch.algos.pearl import PEARLWorker
 from garage.torch.embeddings import MLPEncoder
 from garage.torch.policies import (ContextConditionedPolicy,
                                    TanhGaussianMLPPolicy)
 from garage.torch.q_functions import ContinuousMLPQFunction
-import garage.torch.utils as tu
 from tests.fixtures import snapshot_config
 
 
@@ -108,7 +108,7 @@ class TestPEARL:
             reward_scale=params['reward_scale'],
         )
 
-        tu.set_gpu_mode(params['use_gpu'], gpu_id=0)
+        set_gpu_mode(params['use_gpu'], gpu_id=0)
         if params['use_gpu']:
             pearl.to()
