+++ tests/garage/torch/algos/test_sac.py
@@ -11,10 +11,10 @@ from garage.envs import GarageEnv, normalize
 from garage.experiment import deterministic, LocalRunner
 from garage.replay_buffer import PathBuffer
 from garage.sampler import LocalSampler
+from garage.torch import set_gpu_mode
 from garage.torch.algos import SAC
 from garage.torch.policies import TanhGaussianMLPPolicy
 from garage.torch.q_functions import ContinuousMLPQFunction
-import garage.torch.utils as tu
 from tests.fixtures import snapshot_config
 
 
@@ -209,9 +209,9 @@ def test_sac_inverted_double_pendulum():
               steps_per_epoch=2)
     runner.setup(sac, env, sampler_cls=LocalSampler)
     if torch.cuda.is_available():
-        tu.set_gpu_mode(True)
+        set_gpu_mode(True)
     else:
-        tu.set_gpu_mode(False)
+        set_gpu_mode(False)
     sac.to()
     ret = runner.train(n_epochs=12, batch_size=200, plot=False)
     # check that automatic entropy tuning is used
