+++ tests/garage/torch/optimizers/test_differentiable_sgd.py
@@ -1,8 +1,8 @@
 """Tests for DifferentialSGD optimizer."""
 import torch
 
+from garage.torch import update_module_params
 from garage.torch.optimizers import DifferentiableSGD
-import garage.torch.utils as tu
 
 
 def test_differentiable_sgd():
@@ -20,7 +20,7 @@ def test_differentiable_sgd():
 
     theta_prime = list(policy.parameters())[0]
     loss = torch.sum(theta_prime**2)
-    tu.update_module_params(policy, named_theta)
+    update_module_params(policy, named_theta)
     diff_sgd.zero_grad()
     loss.backward()
